+++ app/src/main/java/com/techcourse/controller/ForwardController.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.asis;
+package com.techcourse.controller;
 
 import context.org.springframework.stereotype.Controller;
 import jakarta.servlet.http.HttpServletRequest;
