+++ app/src/test/java/com/techcourse/support/web/handler/adaptor/HandlerAdaptorsTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
 import org.junit.jupiter.api.Test;
+import webmvc.org.springframework.web.servlet.handler.adaptor.HandlerAdaptors;
 import webmvc.org.springframework.web.servlet.mvc.HandlerAdaptor;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
 import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerAdaptor;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerExecution;
 import java.util.List;
@@ -14,8 +14,7 @@ import java.util.Optional;
 class HandlerAdaptorsTest {
 
     private static final HandlerAdaptors handlerAdaptors = new HandlerAdaptors(
-            List.of(new ManualHandlerAdaptor(),
-                    new AnnotationHandlerAdaptor())
+            List.of(new AnnotationHandlerAdaptor())
     );
 
     @Test
@@ -29,19 +28,5 @@ class HandlerAdaptorsTest {
         //then
         assertThat(optionalHandlerAdaptor).isNotEmpty();
         assertThat(optionalHandlerAdaptor.get()).isInstanceOf(AnnotationHandlerAdaptor.class);
-    }
-
-    @Test
-    void testFindHandlerAdaptorWhenHandlerIsController() {
-        //given
-        final Controller controller = mock(Controller.class);
-
-        //when
-        final Optional<HandlerAdaptor> optionalHandlerAdaptor = handlerAdaptors.findHandlerAdaptor(controller);
-
-        //then
-
-        assertThat(optionalHandlerAdaptor).isNotEmpty();
-        assertThat(optionalHandlerAdaptor.get()).isInstanceOf(ManualHandlerAdaptor.class);
     }
 }
